Refactored edit page onto different url and made buttons navigate between pages
